Matched func_800F65E0_2F6FB0

+++ include/functions.us.h
@@ -344,7 +344,7 @@ void SetBoardFeatureFlag(s32 flag);
 void func_800567D4(void);
 void func_8005E044(s32, s16, u16);
 void func_8005E36C(s16, s32, s32, s32);
-void ExecBoardScene( void(*function), s32);
+void ExecBoardScene( void(*function), s16);
 void func_8006B870(void);
 s32 func_8006B8A4(s32, s32, s32, s32, s32);
 f64 fabs(f64 f);
